# pkg/server/v1/openapi.yaml
openapi: 3.1.0
info:
  title: Peace Breaker Bot API
  version: "1.0"
  description: |
    HTTP API for controlling the Discord bot's playback and for reading
    from myinstants.com. This is the backend for the companion Electron/React
    UI in `peace-breaker-bot-desktop`.

    ## Versioning

    This document describes `v1`, rooted at `/api/v1`. `GET /api/docs`
    renders a human-readable version of it (and, once a v2 exists, a
    switcher between the two); it is deliberately not itself versioned,
    since it documents every version.

    ## Error labels

    `label` is a stable machine-readable code — match on it, not on
    `message` or the HTTP status. `message` defaults to English and
    negotiates via a standard `Accept-Language` header (currently `en-US`
    and `pt-BR` are supported; anything else falls back to English). The
    companion UI ignores `message` entirely and translates `label` itself,
    so it never sends this header. Envelopes distinguish success from error
    by the presence of `data` versus `label`, independent of status code —
    but the status code is now also meaningful and reflects each case below.
servers:
  - url: http://localhost:9001
    description: Default local bind address (server.address / --server-address / SERVER_ADDRESS)
security: []


paths:
  /api/v1/bot/play:
    post:
      operationId: botPlay
      summary: Play a clip through the bot
      description: |
        Resolves `url` to a cached/downloaded clip and starts playing it on
        the bot's current voice connection, stopping whatever is already
        playing first (a bot can only play one clip at a time). Blocks until
        playback ends naturally or is interrupted by `POST /api/v1/bot/stop`,
        then returns which one happened.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BotPlayRequest"
      responses:
        "200":
          description: Playback finished or was interrupted; see `data.exitReason`.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BotPlaySuccessEnvelope"
              examples:
                playedToTheEnd:
                  summary: Playback finished on its own
                  value: { data: { exitReason: end } }
                stopped:
                  summary: Interrupted by POST /api/v1/bot/stop or a newer /api/v1/bot/play
                  value: { data: { exitReason: stop } }
        "400":
          description: The body wasn't valid JSON, or `url` is missing/malformed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
              examples:
                invalidBody:
                  summary: Body was not valid JSON
                  value: { label: invalid_body, message: "Invalid request" }
                invalidUrl:
                  summary: url is missing/malformed
                  value: { label: invalid_url, message: "That URL isn't valid" }
        "404":
          description: The clip could not be resolved.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
              examples:
                instantNotFound:
                  value: { label: instant_not_found, message: "That instant couldn't be found" }
        "409":
          description: The bot has no active voice connection — checked before the clip is resolved or played. See `GET /api/v1/bot/status`.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
              examples:
                botNotConnected:
                  value: { label: bot_not_connected, message: "The bot isn't in a voice channel yet — send !join from the channel you want it in." }
        "422":
          description: The downloaded file did not sniff as mp3.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
              examples:
                unsupportedAudioFormat:
                  value: { label: unsuported_audio_format, message: "That instant isn't an audio format we can play" }
        "500":
          description: Anything else.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
              examples:
                unknownError:
                  value: { label: unknown_error, message: "Something went wrong. Try again in a moment" }

  /api/v1/bot/stop:
    post:
      operationId: botStop
      summary: Stop whatever the bot is currently playing
      description: |
        Signals the in-flight `POST /api/v1/bot/play` (if any) to stop; that
        request then returns with `exitReason: "stop"`. Requests still
        fetching their clip are stopped too. Safe to call when nothing is
        playing (idempotent).
      responses:
        "200":
          description: Empty body; no Content-Type is set. Always this status.

  /api/v1/bot/status:
    get:
      operationId: botStatus
      summary: Report which bot this is and its current voice connection
      description: |
        Read-only, meant for polling: lets the UI pre-emptively disable
        "send to Discord" instead of only reacting to a `409` from
        `POST /api/v1/bot/play`. `bot` identifies the Discord account the
        server is controlling; it is present whether or not the bot is in a
        voice channel, and omitted only until the gateway's `READY` event
        has arrived after startup. Guild/channel/bot IDs are sent as strings —
        Discord snowflakes overflow JS's safe integer range — and the name
        and URL fields are omitted entirely while disconnected, or when
        connected but not found in cache, rather than sent as `""`.
      responses:
        "200":
          description: Always this status; see `data.connected`.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BotStatusEnvelope"
              examples:
                connected:
                  summary: The bot is in a voice channel
                  value:
                    data:
                      connected: true
                      bot:
                        id: "345678901234567890"
                        username: "peace-breaker"
                        displayName: "peace-breaker"
                        avatarUrl: "https://cdn.discordapp.com/avatars/345678901234567890/a1b2c3.png"
                        profileUrl: "https://discord.com/users/345678901234567890"
                        inviteUrl: "https://discord.com/oauth2/authorize?client_id=345678901234567890&scope=bot&permissions=3214336"
                      guildId: "123456789012345678"
                      guildName: "My Server"
                      channelId: "234567890123456789"
                      channelName: "General"
                      channelUrl: "https://discord.com/channels/123456789012345678/234567890123456789"
                disconnected:
                  summary: The bot has no active voice connection
                  value:
                    data:
                      connected: false
                      bot:
                        id: "345678901234567890"
                        username: "peace-breaker"
                        displayName: "peace-breaker"
                        profileUrl: "https://discord.com/users/345678901234567890"
                        inviteUrl: "https://discord.com/oauth2/authorize?client_id=345678901234567890&scope=bot&permissions=3214336"
                gatewayNotReady:
                  summary: The server is up but the gateway hasn't sent READY yet
                  value: { data: { connected: false } }

  /api/v1/instants/{url}/content:
    get:
      operationId: getInstantContent
      summary: Fetch a clip as a base64 data URI
      description: |
        Downloads/caches the clip at `url` (same cache `POST /api/v1/bot/play`
        uses) and returns it inline, without touching the bot or its voice
        connection. Meant for local preview in the UI. `url` is the clip's
        full URL, percent-encoded into this single path segment (e.g. `:`
        becomes `%3A`, `/` becomes `%2F`). `url`'s host must be one of the
        registered providers' declared content hosts (see `GET
        /api/v1/providers`) — a URL that doesn't decode to a usable URL at
        all, and a URL on a host no registered provider recognizes, answer
        the same `invalid_url` 400, so a disallowed host can't be
        distinguished from a malformed one.
      parameters:
        - name: url
          in: path
          required: true
          schema:
            type: string
            format: uri
          description: The clip URL to fetch, percent-encoded. Must be on a host some registered provider serves clips from.
      responses:
        "200":
          description: The clip resolved, or it didn't — see `data.exists`.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InstantContentSuccessEnvelope"
              examples:
                found:
                  summary: Clip resolved and inlined
                  value: { data: { exists: true, content: "data:audio/mp3;base64,//uQx..." } }
                notFound:
                  summary: Clip could not be resolved
                  value: { data: { exists: false } }
        "400":
          description: The path segment didn't decode to a usable URL, or its host isn't one any registered provider allows.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
              examples:
                invalidUrl:
                  value: { label: invalid_url, message: "That URL isn't valid" }
        "422":
          description: The downloaded file did not sniff as mp3.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
              examples:
                unsupportedAudioFormat:
                  value: { label: unsuported_audio_format, message: "That instant isn't an audio format we can play" }
        "500":
          description: Anything else.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
              examples:
                unknownError:
                  value: { label: unknown_error, message: "Something went wrong. Try again in a moment" }
        "502":
          description: url's host answered with a status other than 200/404 (e.g. an access-denied response from the CDN).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
              examples:
                badHttpStatus:
                  value: { label: bad_http_status, message: "That provider's site answered with an error" }

  /api/v1/instants:
    get:
      operationId: listInstants
      summary: List or search a provider's clips
      description: |
        Lists clips from one provider's site — `myinstants` (the default,
        for backward compatibility) or another registered provider key.
        Either browses (no `search`) or searches by name (`search`, which
        ignores `region` — only `myinstants` has a concept of region at
        all; other providers silently ignore it rather than erroring).
        `pages` is inferred from how many instants came back on this page,
        since none of these sites' listing pages carry a real pager: a full
        page means there may be another, a short page means this is the
        last one, and an empty page means the caller paged past the end —
        that last case is normalized into the same collection shape as a
        normal result (`instants: []` plus the last valid `pages`), never a
        bare empty array.
      parameters:
        - name: provider
          in: query
          required: false
          schema:
            type: string
            enum: [myinstants, soundboardguy, soundbuttons]
            default: myinstants
          description: Which registered provider to list from.
        - name: page
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            default: 1
          description: 1-indexed page number. Anything non-positive or non-numeric is treated as 1.
        - name: search
          in: query
          required: false
          schema:
            type: string
          description: Search term. When set on myinstants, region is ignored.
        - name: region
          in: query
          required: false
          schema:
            type: string
            pattern: "^[a-z]{2}$"
            default: us
          description: Two-letter region code, case-insensitive. Only honored by the myinstants provider; other providers ignore it.
      responses:
        "200":
          description: One or more matches, or a search paged past its end.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InstantListSuccessEnvelope"
              examples:
                results:
                  summary: One or more matches
                  value:
                    data:
                      instants:
                        - { name: "Bruh", url: "https://www.myinstants.com/media/sounds/bruh.mp3" }
                      pages: 2
                pastLastPage:
                  summary: A search paged past its end (upstream answered 404)
                  value: { data: { instants: [], pages: 4 } }
        "400":
          description: region did not match ^[a-z]{2}$.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
              examples:
                invalidRegion:
                  value: { label: invalid_region, message: "That region code isn't valid" }
        "404":
          description: provider did not match a registered provider key.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
              examples:
                providerNotFound:
                  value: { label: provider_not_found, message: "That provider isn't recognized" }
        "500":
          description: The scrape found an unexpected markup shape (the site's layout likely changed), or anything else unexpected.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
              examples:
                nameLinkNotMatched:
                  value: { label: name_link_not_matched, message: "That provider's results couldn't be read — its page layout may have changed" }
                unknownError:
                  value: { label: unknown_error, message: "Something went wrong. Try again in a moment" }
        "502":
          description: The provider's site could not be reached, or answered with a bad status.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
              examples:
                httpRequest:
                  summary: Could not reach the provider's site
                  value: { label: http_request, message: "Couldn't reach that provider's site" }
                badHttpStatus:
                  summary: The provider's site answered a non-200/404 status
                  value: { label: bad_http_status, message: "That provider's site answered with an error" }

  /api/v1/providers:
    get:
      operationId: listProviders
      summary: List the registered clip providers
      description: |
        Static-ish list a UI can build a provider picker from, instead of
        hardcoding provider keys client-side. `key` is what `GET
        /api/v1/instants`'s `provider` query param takes. `supportsRegion`
        tells a client whether to show a region field for that provider —
        currently only `myinstants` does anything with it.
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProviderListSuccessEnvelope"
              examples:
                results:
                  value:
                    data:
                      - { key: myinstants, name: MyInstants, supportsSearch: true, supportsRegion: true }
                      - { key: soundboardguy, name: SoundboardGuy, supportsSearch: true, supportsRegion: false }
                      - { key: soundbuttons, name: Sound Buttons, supportsSearch: true, supportsRegion: false }

  /api/v1/openapi.yaml:
    get:
      operationId: getOpenApiSpecV1
      summary: This document
      responses:
        "200":
          description: The v1 OpenAPI document, as YAML.

  /api/docs:
    get:
      operationId: getDocs
      summary: Human-readable rendering of the API's OpenAPI documents
      description: >-
        Not versioned — one docs page for every API version, rendered via
        Redoc with a version switcher above the spec (currently just `v1`).
      responses:
        "200":
          description: An HTML page rendering the available OpenAPI documents.

components:
  schemas:
    Envelope:
      type: object
      description: The shape every response uses, success or error.
      properties:
        label:
          type: string
          description: Present on error responses; a stable machine-readable code.
        message:
          type: string
          description: >-
            Present on error responses; a human-readable message, in
            English by default (see the Error labels section above for
            negotiating a different one via Accept-Language).
        data:
          description: Present on success responses; shape depends on the route.

    ErrorEnvelope:
      allOf:
        - $ref: "#/components/schemas/Envelope"
        - type: object
          required: [label, message]

    BotPlayRequest:
      type: object
      required: [url]
      properties:
        url:
          type: string
          format: uri
          description: The myinstants.com clip URL to play.

    BotPlaySuccessEnvelope:
      allOf:
        - $ref: "#/components/schemas/Envelope"
        - type: object
          required: [data]
          properties:
            data:
              type: object
              required: [exitReason]
              properties:
                exitReason:
                  type: string
                  enum: [end, stop]
                  description: end = playback finished naturally; stop = interrupted.

    BotStatusEnvelope:
      allOf:
        - $ref: "#/components/schemas/Envelope"
        - type: object
          required: [data]
          properties:
            data:
              type: object
              required: [connected]
              properties:
                connected:
                  type: boolean
                bot:
                  type: object
                  description: The Discord account this server controls. Present regardless of `connected`; omitted until the gateway's `READY` event has arrived.
                  required: [id, username, displayName, profileUrl, inviteUrl]
                  properties:
                    id:
                      type: string
                      description: A Discord snowflake, sent as a string.
                    username:
                      type: string
                    displayName:
                      type: string
                      description: The bot's global name if it has one, otherwise its username.
                    avatarUrl:
                      type: string
                      format: uri
                      description: Omitted if no avatar URL could be resolved.
                    profileUrl:
                      type: string
                      format: uri
                      description: Opens the bot's profile in the Discord client.
                    inviteUrl:
                      type: string
                      format: uri
                      description: Add-to-server link with the permissions the bot needs (View Channel, Send Messages, Read Message History, Connect, Speak). Only usable by others if the application is public in Discord's developer portal.
                guildId:
                  type: string
                  description: Present only when connected. A Discord snowflake, sent as a string.
                guildName:
                  type: string
                  description: Present only when connected and resolved from cache.
                channelId:
                  type: string
                  description: Present only when connected. A Discord snowflake, sent as a string.
                channelName:
                  type: string
                  description: Present only when connected and resolved from cache.
                channelUrl:
                  type: string
                  format: uri
                  description: Jump link to the voice channel in the Discord client. Present only when connected.

    InstantContentSuccessEnvelope:
      allOf:
        - $ref: "#/components/schemas/Envelope"
        - type: object
          required: [data]
          properties:
            data:
              type: object
              required: [exists]
              properties:
                exists:
                  type: boolean
                content:
                  type: string
                  description: Only present when exists is true. A base64 data URI, e.g. "data:audio/mp3;base64,...".

    InstantButton:
      type: object
      properties:
        name:
          type: string
        url:
          type: string
          format: uri

    InstantListSuccessEnvelope:
      allOf:
        - $ref: "#/components/schemas/Envelope"
        - type: object
          required: [data]
          properties:
            data:
              type: object
              required: [instants, pages]
              properties:
                instants:
                  type: array
                  items:
                    $ref: "#/components/schemas/InstantButton"
                  description: Always present, even when empty (e.g. a search paged past its end).
                pages:
                  type: integer

    Provider:
      type: object
      properties:
        key:
          type: string
          description: What GET /api/v1/instants's provider query param takes to select this provider.
        name:
          type: string
        supportsSearch:
          type: boolean
        supportsRegion:
          type: boolean
          description: Whether GET /api/v1/instants's region query param does anything for this provider.

    ProviderListSuccessEnvelope:
      allOf:
        - $ref: "#/components/schemas/Envelope"
        - type: object
          required: [data]
          properties:
            data:
              type: array
              items:
                $ref: "#/components/schemas/Provider"